+++ src/main/java/kitchenpos/ordertable/application/dto/TableGroupDto.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.application.dto;
+package kitchenpos.ordertable.application.dto;
 
 import kitchenpos.ordertable.domain.TableGroup;
