+++ src/test/java/kitchenpos/ordertable/application/TableGroupServiceTest.java
@@ -1,17 +1,16 @@
 package kitchenpos.ordertable.application;
 
 import kitchenpos.MockServiceTest;
-import kitchenpos.order.application.dto.CreateOrderTableIdDto;
-import kitchenpos.order.application.dto.CreateTableGroupDto;
-import kitchenpos.order.application.dto.OrderTableDto;
-import kitchenpos.order.application.dto.TableGroupDto;
+import kitchenpos.ordertable.application.dto.CreateOrderTableIdDto;
+import kitchenpos.ordertable.application.dto.CreateTableGroupDto;
+import kitchenpos.ordertable.application.dto.OrderTableDto;
+import kitchenpos.ordertable.application.dto.TableGroupDto;
 import kitchenpos.ordertable.domain.GuestNumber;
 import kitchenpos.order.domain.Order;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.TableGroup;
 import kitchenpos.ordertable.exception.OrderTableException;
 import kitchenpos.ordertable.exception.TableGroupException;
-import kitchenpos.ordertable.application.TableGroupService;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.domain.TableGroupRepository;
 import org.assertj.core.api.Assertions;
